Revert OtlpExporterOptionsBase refactor to shrink the diff of changes.
