+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/ExportClient/BaseOtlpGrpcExportClient.cs
@@ -15,7 +15,7 @@ internal abstract class BaseOtlpGrpcExportClient<TRequest> : IExportClient<TRequ
 {
     protected static readonly ExportClientGrpcResponse SuccessExportResponse = new ExportClientGrpcResponse(success: true, deadlineUtc: null, exception: null);
 
-    protected BaseOtlpGrpcExportClient(OtlpExporterOptionsBase options)
+    protected BaseOtlpGrpcExportClient(OtlpExporterOptions options)
     {
         Guard.ThrowIfNull(options);
         Guard.ThrowIfInvalidTimeout(options.TimeoutMilliseconds);
