+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/Implementation/ExportClient/OtlpGrpcLogExportClient.cs
@@ -11,7 +11,7 @@ internal sealed class OtlpGrpcLogExportClient : BaseOtlpGrpcExportClient<OtlpCol
 {
     private readonly OtlpCollector.LogsService.LogsServiceClient logsClient;
 
-    public OtlpGrpcLogExportClient(OtlpExporterOptionsBase options, OtlpCollector.LogsService.LogsServiceClient logsServiceClient = null)
+    public OtlpGrpcLogExportClient(OtlpExporterOptions options, OtlpCollector.LogsService.LogsServiceClient logsServiceClient = null)
         : base(options)
     {
         if (logsServiceClient != null)
