+++ src/OpenTelemetry.Exporter.OpenTelemetryProtocol/OtlpMetricExporterExtensions.cs
@@ -176,7 +176,7 @@ public static class OtlpMetricExporterExtensions
     }
 
     internal static MetricReader BuildOtlpExporterMetricReader(
-        OtlpExporterOptionsBase exporterOptions,
+        OtlpExporterOptions exporterOptions,
         MetricReaderOptions metricReaderOptions,
         IServiceProvider serviceProvider,
         Func<BaseExporter<Metric>, BaseExporter<Metric>>? configureExporterInstance = null)
